Isolate tests, one per process.
In the absence of Valgrind, this ensures that heap corruption doesn't
spread between tests.  In the presence of Valgrind, this ensures that
error counts and memory leak reports are isolated to the test that
caused them.

It also gives good semantics for any per-process state that tests rely
on, like global variables (e.g. Cyrus config variables).

We start a pipe between the child and parent processes to enable
detailed event reporting, although this is not used yet.