Steps towards a listener-based architecture
Added infrastructure for registering listeners on a globalstate.  If no
listeners get registered, register a default listener which emits
running commentary to stderr.  When forking to run a test, use a special
listener called the proxy listener to capture events and test completion
and push descriptions of them up the event pipe.  This makes the core
code significantly cleaner.  Adding XML and TAP listeners will be a
whole lot easier now.

Report memory leaks and other Valgrind errors as events.  Report
non-zero child exit and child death on a signal as events.