Reorganise state_t internals for >1 exes
External API now has a bool add_executable(const char *filename) instead
of passing the sole filename into the c'tor.  We keep a single unified
vector of all compile units but they point to a new linkobj_t struct
which holds the filename, sections, and mappings.  This will allow
opening multiple bfd objects, which is useful in general but very
necessary when loading all the shared objects for our own running self
(which is the next step).