Reimplement function_t as thin wrapper
instead of converting the DWARF info into data structures in the
function_t when the function_t is created, just store a ref to the dwarf
info and scan it on demand when a caller wants parameter names etc.

This means there's now a large amount of commonality between
function_t::to_string() and the DW_TAG_subroutine_type code in
type_t::to_string, which is not yet used.