Add walker_t::get_path() and move_up()
get_path() returns a vector of references from the enclosing compile
unit down to the current reference.  Currently very slow as it scans the
entire info structure in preorder until it finds the given reference,
even if we got here that way and could have remembered the history.

move_up() moves to the current entry's parent, using get_path().