rel_timestamp epoch is the same across processes
an environment variable is used to track the epoch; if the variable is
missing it's set to the current RT clock, otherwise it's used as the
epoch.

This is tricky because np_init() relies on nobody futzing with the
environment when it discovers the original argv[].

The result is that np::util::rel_timestamp() reports consistent times
across all child processes which actually run tests.