Use executable not object file for dumper tests
The .debug_info section in object files is subject to relocation.  For
32b this doesn't seem to matter, because the addend for relocs which
refer to the string table are actually left in-situ in the .debug_info
section.  But for 64b the .debug_info section contains zeros and the
addend is in the reloc.  This breaks every indirect string attribute.

Fortunately, we don't care about reading debug information from object
files, only from shared libraries and full executables, so we can just
ignore the reloc issue.  Only the tests use an object file, and that's
easily changed.