Fix broken package on SLES11
It turns out that with SLES11's rpmbuild we can't install into
%{_docdir} because the final rpmbuild phase *removes* that directory
just before copying into it anything listed as %doc.  So we have to
use %doc to package the API Reference and Getting Started docs.  But
There are files in doc/get-start/ which we need *not* to package, like
build products from the example, so we need to use a new intermediate
directory doc/inst/.  Oh what a tangled web.