Portability fix.
On Darwin, assert is defined (in assert.h) to be a macro that in turn calls __asert_rtn.  I assume that the intention here is to ensure that calls to assert() are redirected into np, rather than hitting the libc-provided implementation.

Not sure this will work, esp with library namespaces on OSX, but it compiles for now.