Temporarily ifdef out parts of the tintercept test
The tintercept test has some code to test putting intercepts on
functions in libc.  This is important functionality, difficult to
achieve, and needs to be tested.  Unfortunately the functions chosen in
this test program aren't in libc on OSX, they're in a separate -liconv
library.  The test will need to be either reconsidered to be more
portable or made platform specific.  So #ifdef'ed the test out for now.