Handle ASLR when examining live self.
Darwin applies ASLR to the main program by default, which Linux doesn't
do unless you build specially (gcc -fpie).  So to get the stacktrace
test to pass we need to handle ASLR and undo it before looking up
addresses sampled from the stack in DWARF.

The platform-specific layer reports a new 'slide' parameter for each
reported link object.  A range tree of real addresses to linkobj_t is
built in the state_t, which allows the slide to be looked up and used to
undo the mapping.