Intercept leaf or optimized functions in x86
In Ubuntu 14.04 both syslog() and __syslog_chk() in libc are now sufficiently
optimised that they don't start with the "push %ebp" instruction.  Experiment
indicates that only about 30% of functions in libc now start with "push %ebp".
Presumably the Ubuntu folks are building libc with better optimisation.

So the time has come to implement for x86 the 2nd intercept technique earlier
implemented for x86_64.