Re-implement intercepts for Darwin
This turned out to be rather more surgery than I expected.

The hack I was using on Linux to enable setcontext() to return back into
a label in the same function that called setcontext(), turned out not to
work with LLVM/clang, which (once the trace code was added) started
inserting instructions that trashed %rax between the label and the first
__asm__ statement after the label.  This caused lossage of the original
function's return value.

The fix turned out to be fairly complex.  The 2nd half of
intercept_tramp() is now a separate function __np_intercept_after().
The setcontext() call sets up a return address which points at a small
manually-coded assembler trampoline, which saves %rax, discovers the
original stack frame set up by intercept_tramp(), sets up a call to
__np_intercept_after(), and then returns to intercept_tramp()s original
caller.  Actually we need two such routines, one for the "push
%rbp" path and one for the oddball path.

This hasn't been tested with Valgrind yet.