Some junk code for exploring signal delivery
on Darwin with and without Valgrind.  Which turns out to be a mess.

Under Valgrind, SIGCHLD is delivered pseudo-asynchronously by the
Valgrind runtime polling the kernel for pending signals at the
completion of certain system calls.  Valgrind has an async signal
handler for SIGCHLD which is supposed to be enabled during long-running
system calls like poll() to allow the call the return early with EINTR,
but at least in 3.10 this is just broken.  Instead the poll() times out
and the signal is delivered soon afterward (in this code, when the next
sigprocmask() call clears the signal mask).

Valgrind seems to handle this just fine under Linux.  Oh well.