Main loop logic is more resilient
and less dependent on SIGCHLD being delivered in some semblance of good
time.  Explicitly manage the difference between finished and reaped
children in the children_ list, instead of reaping each child as it
finishes and removing it from the list.  This allows us to mix together
children reporting finished (either with an explicit FINISHED message or
by dieing unexpectedly and dropping the pipe) with reaping.  It also
allows reaping to be driven by logic other than delivery of SIGCHLD.  We
do still notice when poll() breaks due to SIGCHLD, and use it to try to
reap early, but no longer rely on this behaviour.

This allows the end of a test to be detected in less than 90 seconds
when running under Valgrind on Darwin.