Better handling for C++ exceptions.
Using a terminate handler is not only easier to write and more portable
but automatically avoids the problem of exceptions which are thrown and
caught in the Code Under Test.

This implementation has a flaw on Linux: the current exception is of
course still allocated when the terminate handler is called, and there's
no visible way to free it, so we end up with spurious Valgrind memory
leak errors which can't even be easily suppressed.  Still, the test was
going to fail anyway.

Also added a unit test.