Better handle corner cases in common path elision.
It turned out that a .c file with no tests but with any of a static
mock, fixture, or parameter, would confuse the common path elision code,
leading to a crash under some circumstances.  The earlier workaround for
this bug in commit 9c88a9e wasn't general enough.  Thanks to Venkatesh
S. for the bug report.