Fix corner case in mapping DWARF sections
In the event that a link object contains DWARF sections and all the
DWARF sections are already mapped by system mappings, the function
np::spiegel::dwarf::state_t::linkobj_t::map_sections() would incorrectly
complain that no sections were found.  Fixed by adjusting the logic.

This does not happen on any platform I have access to, but a user on a
more specialized platform reported a bug consistent with this happening
in the wild.