Show indirect memory leaks.
These leaks are counted towards the reported number of leaks
that we detect and generate an event for, but the default
Valgrind commandline options don't emit them to the memory
leak report, so the test fails and the user can't see why.
Fix is to use a commandline option to show them in the report.