Remove valgrind.supp
NP is designed to be used without needing any Valgrind suppressions
of it's own, and a lot of work has gone into fixing or working around
Valgrind-discovered bugs in NP and in various libraries, and in some
cases feeding back changes to the Valgrind team. Therefore the
valgrind.supp file committed in the repo is redundant. If a suppression
file is required a configure option now exists to supply one.