Re-design reference_t
Instead of containing an index into the state_t's tables of either
compile units or link objects, reference_t now points at those directly.
This allows a reference to be resolved without finding a state_t, which
removes most of the remaining need for state_t::instance().

The downside is that link_object_t now has an up-pointer to state_t.
This is a small price to pay.