Fix occasional incorrect compile units in stack traces
The DWARF v3 standard says

> Debugging information entries that represent a declaration that
> completes another (earlier) non- defining declaration, may have a
> DW_AT_specification attribute whose value is a reference to the
> debugging information entry representing the non-defining declaration.
> Debugging information entries with a DW_AT_specification attribute do
> not need to duplicate information provided by the debugging information
> entry referenced by that specification attribute.

NovaProva has for some years incorrectly implemented doing an
address-to-function lookup on functions whose debugging entries did
this, with the effect that such functions would not be entered in the
address range tree, ultimately resulting in bad guesswork being applied
to figuring out which compile unit a function resides in.  This happens
for both gcc and clang when a class member function is declared and
defined separately, like:

// succulents.hxx
class succulents_t
{
    unsigned long normcore();
};

// tdescaddr.cxx
unsigned long succulents_t::normcore()
{ ... }

One such is, ironically enough, np::spiegel::describe_stacktrace() which
is called to describe it's own address when a stack trace is printed.
The tstack unit test was triggering this bug but I had stupidly covered
it up in the golden output; the unit test failed on Mac because NP took
a different bad guess for the compile unit than on Linux.

Fixes:

- do not dereference DW_AT_specification when building the address
  range tree.  The address range information is in the DIE which
  describes the definition, which has no name or type information
  but does have a DW_AT_specification pointing to another DIE
  which describes the declaration.

- adjusted golden output for tstack test

- added a more rigorous test tdescaddr