Refactor how intercepts are installed
Intercepts broke on Linux as a side effect of commit

040918d Fix intercepts on MacOS Catalina

After that commit, the page containing the intercept address was left
with protection rw- between the calls to text_map_writable() and
text_restore().  It turns out that due to an accident of ELF executable
layout on Linux, it's possible for this to result in a PLT which cannot
be executed.  This is turn causes all calls to functions in shared
libraries (including libc and libstd++) to SEGV, including

- the sigaction() call which installs the signal handler
- the map::erase function that was decrementing page refcounts
- the mprotect() call which restores protections
- any debugging prints

Also, intercepts of functions whose first instruction is not "push %rbp"
need to temporarily uninstall and reinstall the trap code from the tramp
function, which was easy when the text page remained writable but is no
longer easy.

The fix is a fairly hefty refactor of the intercept installation code.

- Move code around to avoid or eliminate all calls via the PLT between
  the two calls to mprotect().  In particular we use some RTLD magic to
  get the real address of libc's mprotect() and call that directly.
- Make text_map_writable() and text_restore() private and wrap them
  in a new text_write() function which also does the memory copying
  and invalidates Valgrind translations.  This puts the entire "PLT
  might be broken" critical section into one nicely contained and
  documented function.
- Call text_write() from all 4 places where the bytes in the text need
  modifying: install_intercept(), uninstall_intercept(), and twice in
  the tramp function.
- Eliminate the now-useless page reference counting scheme.