Intercepts preserve ABI mandated stack alignment on Linux x86-64
When a function whose function prolog does not start with the
common "push %rbp" sequence was intercepted, then either the
original function or a redirect function would be called with a mis-
aligned stack.  More precisely, after the called function's prolog
finished, it's RSP and RBP would be 8 modulo 16.

The SysV ABI for x86-64 mandates a 16B alignment for the stack, so
NovaProva was always incorrect.  However we got away with it for a long
time because none of the called functions actually cared that much about
stack alignment, thanks to the architecture's historical insensitivity
to alignment of memory accesses.

This changed sometime between Ubuntu Trusty and Ubuntu Bionic,
presumably as a result of compiler improvements.  The regexec()
function in libc now uses the MOVAPS instruction from SSE to efficiently
zero out a 32B structure on the stack, and that instruction results in a
SEGV if the memory access is not 16B aligned.  That's a perfectly valid
thing to do, but NovaProva's misaligning behavior broke it.  This was
causing the tnsyslogmatch unit test to fail.

The fix is to simply insert an extra word of padding in the artifical
stack frame fed by the tramp to the original & redirect functions, for
the OTHER intercept type.

This bug dates back to the original x86-64 implementation in 2012.