Cleaner fix for Linux x86_64 stack alignment
Commit 21a7339 fixed stack alignment of original and redirect functions
called from the Linux x86-64 intercept tramp.  It turns out this fix
works when built with gcc but not with clang, as it implicitly relied on
the alignment of a stack variable.

This new fix ports across to Linux x86-64 the similar but cleaner fix to
Darwin x86-64 from commit 5108e4c.  That fix does several things more
gracefully but in particular forces alignment of the "frame" struct with
an __attribute__.  Tests indicate it works on both compilers.