Fix mprotect() collateral damage again, more generally
It turns out that the problem of an accidentally unexecutable PLT
fixed in 1cb210d is a little more general.  It's possible for
a function in libc which is intercepted to lie on the same text
page as the mprotect() system call stub.  This causes the mprotect()
call to succeed and to accidentally remove execute permission from
the instruction following the trap into the kernel, resulting in
a SEGV looking like this

==20374==  Bad permissions for mapped region at address 0x5FDDEB7
==20374==    at 0x5FDDEB7: mprotect+7 (syscall-template.S:84)
==20374==    by 0x4453D0: np::spiegel::platform::mprotect(void*, unsigned long, int)+77 (mprotect.cxx:42)

The fix is to change from bypassing the PLT but using the real libc
mprotect(), to just making our own copy of mprotect() from the static
libc with alignment tricks to ensure it does not share a page with
any other code at all.  This solves both flavors of the problem.