Fix skip() on OTHER type function intercepts
A logic bug in the platform-specific intercept trampolines meant that if
an intercept's before() method is called and decides to skip() calling
the original function, and the intercept was installed as type OTHER
(because the intercepted function did not start with the common "push
%rbp" instruction), then the intercept trap code was not properly
re-inserted.  This has the effect that the intercept was effectively
uninstalled by accident.

Discovered on Ubuntu 20.04, where the default prologue always starts
with the endbr64 instruction.  However this is a very long-standing bug,
dating back to the very first implementation from which all the
platform-specific trampolines derive.  It shows a gap in the unit
testing of intercepts.

Also added some x86-64 specific intercept tests which explicitly test
various cases of function prologues, and a new method in intercept_t
which allows for checking whether an intercept is actually installed (it
checks the first instruction in the function)