Fix intercept of getpid() et al on Darwin
It turns out the reason that we couldn't intercept some syscall stubs
like getpid() on Darwin was that the system call stub was accidentally
located on the same page as the stub for vm_protect(), so the first
vm_protect() call will accidentally make itself unexecutable.

This problem was already solved on Linux in commit
5ee6a41009186f5c040a8822f406503b6bfd263b so the fix is to apply the same
trick, namely a manually crafted copy of the vm_protect() system call
stub which is carefully organized to be on a .text page by itself.

Update the tintercept test to test intercepting getpid() et al on all
platforms.