Fix tintercept "cardigan" test case on Darwin
Two separate issues:

- One of the four locations where the instruction stream is modified to
  add or remove a breakpoint, was not updated to call text_write() when
  that function was added in commit 1cb210d.  The broken code was
  attempting to write to the intercepted location directly and getting a
  SEGV.  Fixed by calling text_write().

- One of the two code paths for how intercepted functions are called was
  not correctly adjusted to cope with the way that x86_65 stack
  alignment was enforced in commit 5108e4c.  By sheer accident this was
  resulting in a stack address being passed to text_write() when an
  OTHER function was intercepted, and a page of the stack being marked
  r-x, which needless to say resulted in a rapid SEGV.  Fixed by
  changing the magic stack offset in the hand-rolled assembler.

Also added a bunch of trace-level debugging.

Also added some defensive programming.  On Darwin the fake stack frame
is manipulated in two separate functions, and its super critical that
the address is correctly discovered in the second one.  Added a 4-byte
magic number in the structure, and checked it in the second function
to ensure a relatively graceful failure mode if it goes wrong again.

All this was uncovered by the "cardigan" test case in the tintercept
test, which reliably forces the OTHER mode of interception.