Work around compilers which default to DWARF 5
The gcc compiler in some modern distros defaults to DWARF version 5 which
NovaProva doesn't support yet.  This default results in a library and test
programs built in such a way that NP cannot parse their debug information,
which in turn leads to all of NP's tests failing.  As a short term measure,
detect and handle this case in the configure script by adding a platform
compile flag to force the highest version that NP supports.