GHA CI move all the CI shell code into separate scripts
located in build/ci/.  This allows that setup logic to be
reused for other environments.

Add a build/docker.sh which runs either
- a full build&test cycle, or
- a prep cycle and starts an interactive shell
in a new container based on the given image.  This allows
debugging build problems locally.

Used that to diagnose the Fedora 34 problem, and documented
it in a comment for now.

Add some gitignores.