Gracefully handle failure to read DWARF compile unit headers
Instead of exiting the program, emit a useful error or warning message and skip
to the next compile unit in the same .info section.  Unexpected DWARF version
is now a warning not an error.

This is necessary in order to cope with compile units from Fedora 34's libbfd.a
which is built with DWARF5 debug info.